refactor(search-react): apply review suggestions
